Support an OpenURLs event on macOS
This hooks into the application:openUrls delegate callback on macOS to deliver
a corresponding `Event::OpenURLs` to applications.

Although this event is currently only supported on macOS; potentially in the
future this could also be supported on Linux by handling the
org.freedesktop.Application.Open method and potentially Android Intent's with a
URL could be mapped to OpenURLs events via some Activity::onNewIntent
integration.