Workaround for WM_PAINT buffering conflicting with MessageBoxW
To allow better interoperability with other crates that directly
interact with Window messages the Windows event_loop needs a different
solution for buffering redraws that can defer the RedrawWindow call
until just before we poll for messages.

We're using the `rfd` crate for native message boxes and that conflicts
with Winit because APIs like `MessageBoxW` seem to try and flush pending
redraws before starting to draw new modal dialogs.

For now, we're assuming that our application redraws continuously
and so it won't break if we skip buffering these redraw requests.

Arguably this isn't really an issue with Winit, so it probably doesn't
make sense to try and upstream this change.

I tend to think we're asking for trouble by by using the `rfd` crate
which expects to be able to interact with the window system without any
integration with Winit and should either find a way to show dialogs in a
way that's better integrated with Winit or avoid trying to show native
dialogs.